refactor:(LibroResumeDTO): change type price DOUBLE to BigDecimal

+++ src/main/java/com/io/librotech/dto/LibroResumeDTO.java
@@ -1,6 +1,7 @@
 package com.io.librotech.dto;
 
 import java.time.LocalDate;
+import java.math.BigDecimal;
 
-public record LibroResumeDTO(Long id, String titulo, LocalDate fechaPublicacion, Double precio, String editorialNombre, String pais) {
+public record LibroResumeDTO(Long id, String titulo, LocalDate fechaPublicacion, BigDecimal precio, String editorialNombre, String pais) {
 }
